+++ maat-court-data-api/src/main/java/gov/uk/courtdata/enums/HardshipReviewDetailType.java
@@ -2,8 +2,7 @@ package gov.uk.courtdata.enums;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonValue;
-import gov.uk.courtdata.helper.AbstractEnumConverter;
-import gov.uk.courtdata.helper.PersistableEnum;
+import gov.uk.courtdata.converter.AbstractEnumConverter;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
